Renamed box names
