+++ src/routes/levels/[slug]/Goal.svelte
